project completed

+++ backend/controller/product.controller.js
@@ -49,11 +49,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "product id invalid" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "product deleted successfully." });
     } catch (error) {
         console.error("Error while deleting :" + error)
-        res.status(404).json({ success: false, message: "product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
